requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
yarg==0.1.9
